requirements.gui.txt
Flask>=3.0.0
requests>=2.32.3
python-dotenv>=1.0.1

